fix type -Wconvertion

+++ src/hal/board/arduino_mega/hal_ZS_042.c
@@ -21,7 +21,7 @@
 
 #define ZS_042_I2C_ADDR 0x68
 
-static uint8_t bcdToUint8(uint8_t bcd) { return ((bcd >> 4) * 10u) + (bcd & 0x0Fu); }
+static uint8_t bcdToUint8(uint8_t bcd) { return (uint8_t)((bcd >> 4) * 10u) + (bcd & 0x0Fu); }
 
 static uint8_t uint8ToBcd(uint8_t val) { return ((val / 10u) << 4) | (val % 10u); }
